撮合bug

+++ token_service/model/trade_model.go
@@ -11,7 +11,7 @@ const (
 type Trade struct {
 	TradeId      int    `xorm:"not null pk comment('交易表的id') INT(11)"`
 	TradeNo      string `xorm:"comment('订单号') VARCHAR(32)"`
-	Uid          int    `xorm:"comment('买家uid') index INT(11)"`
+	Uid          uint64 `xorm:"comment('买家uid') index INT(11)"`
 	TokenId      int    `xorm:"comment('主货币id') index INT(11)"`
 	TokenTradeId int    `xorm:"comment('交易币种') INT(11)"`
 	Price        int64  `xorm:"comment('价格') BIGINT(20)"`
